ロジックをカスタムフックにリファクタリング
